Add some testing files
